promptclass-app/requirements.txt
flask
torch
torchvision
transformers
pillow
flask-cors
